Move 'map_unary_minus_int_func' to rpc_expr/mod.rs
